Game 1 Week In
Needs work on player, needs better jumping, ducking and there is still
no spinning.  Also we should research a player model instead of a block.

We need power ups for different parts of the songs along with screen
shakes and other cool effects.

The platform (floor) moving is the current method of keeping all the
game objects moving in the screen.